+++ pkg/http/jwt/handler.go
@@ -20,7 +20,6 @@ type Credentials struct {
 
 func CreateNewTokens() (authTokenString, refreshTokenString, csrfSecret string, err error) {
 	csrfSecret = CSRFKey
-	refreshTokenString, err = createRefreshTokenString(csrfSecret)
 	authTokenString, err = createAuthTokenString(csrfSecret)
 
 	if err != nil {
@@ -29,7 +28,7 @@ func CreateNewTokens() (authTokenString, refreshTokenString, csrfSecret string, 
 	return
 }
 
-func SetCookies(write http.ResponseWriter, authTokenString string, refreshTokenString string) {
+func SetCookies(write http.ResponseWriter, authTokenString string) {
 	authCookie := http.Cookie{
 		Name:     "AuthToken",
 		Value:    authTokenString,
@@ -37,14 +36,6 @@ func SetCookies(write http.ResponseWriter, authTokenString string, refreshTokenS
 		Path:     "/",
 	}
 	http.SetCookie(write, &authCookie)
-
-	refreshCookie := http.Cookie{
-		Name:     "RefreshToken",
-		Value:    refreshTokenString,
-		HttpOnly: true,
-		Path:     "/",
-	}
-	http.SetCookie(write, &refreshCookie)
 }
 
 func createAuthTokenString(csrfSecret string) (authTokenString string, err error) {
@@ -63,22 +54,6 @@ func createAuthTokenString(csrfSecret string) (authTokenString string, err error
 	return
 }
 
-func createRefreshTokenString(csrfString string) (refreshTokenString string, err error) {
-	expirationTime := time.Now().Add(72 * time.Hour).Unix()
-
-	refreshClaims := &Claims{
-		CSRF: csrfString,
-		StandardClaims: jwt.StandardClaims{
-			NotBefore: time.Now().Unix(),
-			ExpiresAt: expirationTime,
-			IssuedAt:  time.Now().Unix(),
-		},
-	}
-	refreshJwt := jwt.NewWithClaims(jwt.SigningMethodHS256, refreshClaims)
-	refreshTokenString, err = refreshJwt.SignedString(JwtKey)
-	return
-}
-
 func ClearSession(write http.ResponseWriter) {
 	authCookie := http.Cookie{
 		Name:     "AuthToken",
@@ -88,31 +63,4 @@ func ClearSession(write http.ResponseWriter) {
 		Path:     "/",
 	}
 	http.SetCookie(write, &authCookie)
-
-	refreshCookie := http.Cookie{
-		Name:     "RefreshToken",
-		Value:    "",
-		Expires:  time.Now().Add(-1000 * time.Hour),
-		HttpOnly: true,
-		Path:     "/",
-	}
-	http.SetCookie(write, &refreshCookie)
-}
-
-func getandrefreshtokens() {
-	// Verifier la validité de mon cookie
-	// si c'est valid : Return
-	//Si pas valid
-	// recupere le refresh dans ma bdd
-	// si c'est valide je refresh le AuthToken
-	// Sinon return 401
-}
-
-func GetCsrfFromReq(r *http.Request) string {
-	csrfFromForm := r.FormValue("X-CSRF-Token")
-	if csrfFromForm != "" {
-		return csrfFromForm
-	} else {
-		return r.Header.Get("X-CSRF-Token")
-	}
 }
